requirements.txt
asgiref==3.8.1
Django==5.1
django-crispy-forms==2.3
django-haystack==3.3.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
Whoosh==2.7.4